Upload userContext.js
At this time I cannot fully recall what this file is doing but it is important to be in here in order to get around a react error when running. It's been about 8-9 months since I've worked on this project and am just trying to get it to a stable state to work and be presentable right now.